+++ my_vrg/Views/MostSharedView.swift
@@ -42,7 +42,6 @@ struct MostSharedView: View {
                         }
                     }
                     .listStyle(PlainListStyle())
-
                 }
             }
             .navigationTitle("Most shared")
